+++ tests/test_image.py
@@ -1,9 +1,14 @@
-import requests
+import base64
+from pathlib import Path
 from io import BytesIO
 
+import requests
+
 from PIL import Image
-from pyturbocode.TurboCodec import TurboCodec
-from imhide.crypto import generate_key, encrypt
+import pypccc
+from struct import pack, unpack
+
+from imhide.crypto import decrypt, generate_key, encrypt
 
 
 def _load_image() -> bytes:
@@ -11,24 +16,59 @@ def _load_image() -> bytes:
     response = requests.get(url)
     image_data = BytesIO(response.content)
 
-    # with open("output.jpg", "wb") as f:
-    #     f.write(image_data.getbuffer())
-
     image = Image.open(image_data).convert("RGBA")
     data = image.tobytes()
 
     return data
 
 
-def test_image():
-    key = generate_key()
-    codec = TurboCodec()
+def create_encoded_image(key: bytes, data: bytes, encoded_file_pth: Path):
+    # (n=bin_token size)
+    # n	K	R	C	C2	p	q	ratio
+    # 3162914	4	1.461538462	3724	13868176	1862	1274	1.00543995
 
-    data = _load_image()
-
+    p = 1862
+    q = 1274
     token = encrypt(data, key)
-    enc = codec.encode(token)
-    print(len(enc))
+    bin_token = base64.urlsafe_b64decode(token)
+    n_tkn = len(bin_token)
+    n_pad = 3162914 - n_tkn - 8
+    bin_token = pack("Q", n_tkn) + bin_token + b"\x00" * n_pad
+    enc = pypccc.rs_encode(bin_token)
+
+    image = Image.frombytes(mode="RGBA", data=enc, size=(p, q))
+
+    with open(encoded_file_pth, "wb") as f:
+        image.save(f)
+
+
+def get_decoded_image(key: bytes, encoded_file_pth: Path):
+    with open(encoded_file_pth, "rb") as f:
+        image = Image.open(f).convert("RGBA")
+    enc = image.tobytes()
+    bin_token = pypccc.rs_decode(enc)
+    n_tkn = unpack("Q", bin_token[:8])[0]
+    token = base64.urlsafe_b64encode(bin_token[8 : 8 + n_tkn])
+    decoded_data = decrypt(token, key)
+    return decoded_data
+
+
+def test_image():
+    img_pth = Path("output.png")
+    new_img_pth = Path("decoded.png")
+    key = generate_key()
+
+    original = _load_image()
+    create_encoded_image(key, original, encoded_file_pth=img_pth)
+    print("Encoded image created")
+
+    decoded = get_decoded_image(key, img_pth)
+
+    image = Image.fromarray(decoded.astype("uint8"), "RGB")
+
+    with open(new_img_pth, "wb") as f:
+        image.save(f)
+    print("Decoded image created")
 
 
 if __name__ == "__main__":